refactor: replace localforage with localStorage for cookie storage
Drop the localforage dependency in favor of plain localStorage for
cookie settings. Inline the simplified cookie validation logic into
cookies.ts (removed cookieFormat.ts). Simplify the admin cookies API
route by removing cookie format canonicalization and validation.
