+++ src/app/api/admin/cookies/route.ts
@@ -2,7 +2,6 @@ import { NextResponse } from "next/server";
 import { promises as fs } from "fs";
 import { existsSync } from "fs";
 import path from "path";
-import { canonicalizeCookiesContent, validateCookiesContent } from "@/lib/cookieFormat";
 
 const ADMIN_PASSWORD = process.env.ADMIN_PASSWORD;
 const DATA_DIR = path.join(process.cwd(), "data");
@@ -10,30 +9,24 @@ const SYSTEM_COOKIES_PATH = path.join(DATA_DIR, "cookies.txt");
 
 function verifyPassword(request: Request): boolean {
   const authHeader = request.headers.get("Authorization");
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    return false;
-  }
-  const password = authHeader.slice(7);
-  return password === ADMIN_PASSWORD;
+  if (!authHeader || !authHeader.startsWith("Bearer ")) return false;
+  return authHeader.slice(7) === ADMIN_PASSWORD;
 }
 
 export async function GET(request: Request) {
   if (!ADMIN_PASSWORD) {
     return NextResponse.json({ error: "Admin not configured" }, { status: 403 });
   }
-
   if (!verifyPassword(request)) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
 
   try {
-    if (existsSync(SYSTEM_COOKIES_PATH)) {
-      const cookies = await fs.readFile(SYSTEM_COOKIES_PATH, "utf-8");
-      return NextResponse.json({ cookies });
-    }
-    return NextResponse.json({ cookies: "" });
-  } catch (error) {
-    console.error("[Moonlit] Error reading system cookies:", error);
+    const cookies = existsSync(SYSTEM_COOKIES_PATH)
+      ? await fs.readFile(SYSTEM_COOKIES_PATH, "utf-8")
+      : "";
+    return NextResponse.json({ cookies });
+  } catch {
     return NextResponse.json({ error: "Failed to read cookies" }, { status: 500 });
   }
 }
@@ -42,38 +35,18 @@ export async function POST(request: Request) {
   if (!ADMIN_PASSWORD) {
     return NextResponse.json({ error: "Admin not configured" }, { status: 403 });
   }
-
   if (!verifyPassword(request)) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
 
   try {
     const { cookies } = await request.json();
-
     if (!existsSync(DATA_DIR)) {
       await fs.mkdir(DATA_DIR, { recursive: true });
     }
-
-    if (!cookies || !cookies.trim()) {
-      await fs.writeFile(SYSTEM_COOKIES_PATH, "", "utf-8");
-    } else {
-      const validation = validateCookiesContent(cookies);
-      if (!validation.valid) {
-        return NextResponse.json(
-          { error: validation.error || "Invalid cookie format" },
-          { status: 400 },
-        );
-      }
-      await fs.writeFile(
-        SYSTEM_COOKIES_PATH,
-        canonicalizeCookiesContent(cookies),
-        "utf-8",
-      );
-    }
-
+    await fs.writeFile(SYSTEM_COOKIES_PATH, cookies || "", "utf-8");
     return NextResponse.json({ success: true });
-  } catch (error) {
-    console.error("[Moonlit] Error writing system cookies:", error);
+  } catch {
     return NextResponse.json({ error: "Failed to save cookies" }, { status: 500 });
   }
 }
