--- src/lib/cookieFormat.ts
@@ -1,122 +1,0 @@
-const MOZ_COOKIE_HEADER_LINES = [
-  "# Netscape HTTP Cookie File",
-  "# https://curl.se/rfc/cookie_spec.html",
-];
-
-/** True if line is a recognizable Mozilla/Netscape cookie header comment. */
-function looksLikeMozCookieHeader(line: string): boolean {
-  const t = line.trim();
-  return (
-    /^#\s*(Netscape )?HTTP Cookie File/i.test(t) ||
-    t === "# HTTP Cookie File" ||
-    /^#\s+This file is generated by yt-dlp/i.test(t)
-  );
-}
-
-/**
- * Split one Netscape data line into 7 fields (domain, subdomain, path, secure, expiry,
- * name, value). Tabs are canonical; pasted text often collapses tabs into spaces.
- */
-function splitNetscapeCookieLine(line: string): string[] | null {
-  const trimmed = line.trim();
-  if (!trimmed || trimmed.startsWith("#")) return null;
-
-  const byTab = trimmed.split("\t");
-  if (byTab.length >= 7) {
-    const [domain, sub, path, secure, expiry, name, ...valueParts] = byTab as [
-      string,
-      string,
-      string,
-      string,
-      string,
-      string,
-      ...string[],
-    ];
-    return [
-      domain,
-      sub || "",
-      path || "",
-      secure || "",
-      expiry || "",
-      name || "",
-      valueParts.join("\t"),
-    ];
-  }
-
-  const m = trimmed.match(
-    /^(\S+)\s+(TRUE|FALSE)\s+(\S+)\s+(TRUE|FALSE)\s+(\S+)\s+(\S+)\s*(.*)$/i,
-  );
-  if (!m) return null;
-  const sub = /^true$/i.test(m[2]!) ? "TRUE" : "FALSE";
-  const secure = /^true$/i.test(m[4]!) ? "TRUE" : "FALSE";
-  return [m[1]!, sub, m[3]!, secure, m[5]!, m[6]!, m[7] ?? ""];
-}
-
-/** Rewrite pasted cookies into strict tab-separated Netscape format yt-dlp expects. */
-export function canonicalizeCookiesContent(content: string): string {
-  const normalized = content.replace(/\r\n/g, "\n").replace(/\r/g, "\n");
-  const rawLines = normalized.split("\n");
-  let hasMozHeader = false;
-  const outLines: string[] = [];
-
-  for (const raw of rawLines) {
-    const trimmedEnd = raw.trimEnd();
-    if (looksLikeMozCookieHeader(trimmedEnd)) hasMozHeader = true;
-    const trimmedLeft = raw.trimStart();
-    if (!trimmedLeft || trimmedLeft.startsWith("#")) {
-      outLines.push(raw);
-      continue;
-    }
-    const parts = splitNetscapeCookieLine(trimmedLeft);
-    if (parts) outLines.push(parts.join("\t"));
-    else outLines.push(raw);
-  }
-
-  let dataRowCount = 0;
-  for (const l of outLines) {
-    if (splitNetscapeCookieLine(l.trim())) dataRowCount += 1;
-  }
-
-  const lead: string[] = [];
-  if (dataRowCount > 0 && !hasMozHeader) {
-    for (const h of MOZ_COOKIE_HEADER_LINES) lead.push(h);
-    lead.push("");
-  }
-
-  return [...lead, ...outLines].join("\n").replace(/\n+$/, "").concat("\n");
-}
-
-/** Validate Netscape cookie format (tabs or pasted space-separated cookie rows). */
-export function validateCookiesContent(content: string): {
-  valid: boolean;
-  error?: string;
-} {
-  if (!content.trim()) return { valid: true };
-
-  const working = canonicalizeCookiesContent(content);
-  const lines = working.split("\n");
-  let seenDataRow = false;
-
-  for (const line of lines) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-
-    if (!splitNetscapeCookieLine(trimmed)) {
-      return {
-        valid: false,
-        error: `Invalid cookie line (need 7 fields: domain through value): "${trimmed.substring(0, 80)}${trimmed.length > 80 ? "..." : ""}"`,
-      };
-    }
-    seenDataRow = true;
-  }
-
-  if (!seenDataRow && content.trim().length > 0) {
-    return {
-      valid: false,
-      error:
-        "No valid cookie rows found - paste Netscape-format rows or a full cookies.txt.",
-    };
-  }
-
-  return { valid: true };
-}
