+++ src/lib/cookies.ts
@@ -1,34 +1,28 @@
-import localforage from "localforage";
-import { canonicalizeCookiesContent, validateCookiesContent } from "@/lib/cookieFormat";
+const COOKIES_KEY = "moonlit-yt-cookies";
+const CUSTOM_COOKIES_ENABLED_KEY = "moonlit-custom-cookies-enabled";
 
-const store = localforage.createInstance({
-  name: "moonlit",
-  storeName: "settings",
-});
-
-const COOKIES_KEY = "yt-cookies";
-const CUSTOM_COOKIES_ENABLED_KEY = "custom-cookies-enabled";
-
-/** Get user cookies from local storage */
+/** Get user cookies from localStorage */
 export async function getUserCookies(): Promise<string> {
-  const value = await store.getItem<string>(COOKIES_KEY);
-  return value ?? "";
+  if (typeof window === "undefined") return "";
+  return localStorage.getItem(COOKIES_KEY) ?? "";
 }
 
-/** Save user cookies to local storage (canonical tab-separated Moz format) */
+/** Save user cookies to localStorage */
 export async function setUserCookies(content: string): Promise<void> {
-  await store.setItem(COOKIES_KEY, canonicalizeCookiesContent(content));
+  if (typeof window === "undefined") return;
+  localStorage.setItem(COOKIES_KEY, content);
 }
 
 /** Get custom cookies enabled preference */
 export async function isCustomCookiesEnabled(): Promise<boolean> {
-  const value = await store.getItem<boolean>(CUSTOM_COOKIES_ENABLED_KEY);
-  return value === true;
+  if (typeof window === "undefined") return false;
+  return localStorage.getItem(CUSTOM_COOKIES_ENABLED_KEY) === "true";
 }
 
 /** Set custom cookies enabled preference */
 export async function setCustomCookiesEnabled(enabled: boolean): Promise<void> {
-  await store.setItem(CUSTOM_COOKIES_ENABLED_KEY, enabled);
+  if (typeof window === "undefined") return;
+  localStorage.setItem(CUSTOM_COOKIES_ENABLED_KEY, String(enabled));
 }
 
 /** Get cookies to use for requests (user cookies if enabled, otherwise empty) */
@@ -43,7 +37,33 @@ export async function getCookiesToUse(): Promise<{ cookies: string }> {
   return { cookies: "" };
 }
 
-/** Validate Netscape cookie format (tabs or pasted space-separated cookie rows) */
+/** Validate Netscape cookie format — each non-empty, non-comment line must have 7 tab-separated fields */
 export function validateCookies(content: string): { valid: boolean; error?: string } {
-  return validateCookiesContent(content);
+  if (!content.trim()) return { valid: true };
+
+  const lines = content.split("\n");
+  let seenDataRow = false;
+
+  for (const line of lines) {
+    const trimmed = line.trim();
+    if (!trimmed || trimmed.startsWith("#")) continue;
+
+    const parts = trimmed.split("\t");
+    if (parts.length < 7) {
+      return {
+        valid: false,
+        error: `Each cookie line needs 7 tab-separated fields. Got ${parts.length}.`,
+      };
+    }
+    seenDataRow = true;
+  }
+
+  if (!seenDataRow) {
+    return {
+      valid: false,
+      error: "No valid cookie rows found.",
+    };
+  }
+
+  return { valid: true };
 }
